third_party/blink/web_tests/storage/quota/storagequota-query-usage-expected.txt

This tests querying usage and quota using Quota API.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS returnedUsage is >= 0
PASS returnedQuota is >= returnedUsage
PASS successfullyParsed is true

TEST COMPLETE
